RegionOfInterestProcess supporting camera offset option
