+++ src/main/java/org/openpnp/gui/processes/RegionOfInterestProcess.java
@@ -29,7 +29,9 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import javax.swing.SwingUtilities;
@@ -38,9 +40,12 @@ import org.openpnp.gui.MainFrame;
 import org.openpnp.gui.components.CameraView;
 import org.openpnp.gui.components.reticle.Reticle;
 import org.openpnp.model.LengthUnit;
+import org.openpnp.model.Location;
 import org.openpnp.model.RegionOfInterest;
+import org.openpnp.model.RegionOfInterestOffset;
 import org.openpnp.spi.Camera;
 import org.openpnp.util.UiUtils;
+
 
 /**
  * Guides the user through the three point region of interest operation using step by step instructions.
@@ -51,27 +56,54 @@ public class RegionOfInterestProcess {
     private final Camera camera;
     private final CameraView cameraView; 
 
-    private int step = -1;
     private String[] instructions = new String[] {
+            "<html><body>Camera position can be offset with jog or by selecting position in the camera view. Click Next to continue.</body></html>",
             "<html><body>Click on what will become the upper left corner of your region of interest. Click again to reset and retry. Click Next to continue.</body></html>",
             "<html><body>Now, click on what will become the upper right corner of your region of interest. Click again to reset and retry. Click Next to continue.</body></html>",
             "<html><body>Next, click on what will become the lower left corner of your region of interest. Click again to reset and retry. Click Next to continue.</body></html>",
             "<html><body>Finally, click to toggle beween a rectangular and parallelogrammatic region. Click Next to continue.</body></html>",
-            "<html><body>The region of interest has been defined. Click Finish to accept it, or Cancel to quit.</body></html>",};
+            "<html><body>The region of interest has been defined. Click Finish to accept it, or Cancel to quit.</body></html>"};
+    
+    private enum ROIStep {
+        Init,
+        CameraOffset,
+        UpperLeft,
+        UpperRight,
+        LowerLeft,
+        SelectMode,
+        Complete,
+        Save
+    };
+    
+    final List<ROIStep> stepSequence = Arrays.asList(ROIStep.values());
+    
 
-    private Map<Integer, Point> regionStakeout = new HashMap<>();
+    private int stepIndex = 0;
+    private ROIStep step =  ROIStep.Init;
+
+    private Map<ROIStep, Point> regionStakeout = new HashMap<>();
     private boolean rectify = false;
     private Point mouseLastPos = null;
     private int mouseClickCount = 0;
 
     private static final String PROCESS_RETICLE_KEY = "PROCESS_RETICLE_KEY";
 
-    RegionOfInterest regionOfInterest = null;
+    private RegionOfInterest regionOfInterest = null;
+    private RegionOfInterestOffset regionOfInterestOffset= null;
+    private boolean doLocation = false;
+    private Location startLocation = null;
+    
 
-    public RegionOfInterestProcess(MainFrame mainFrame, Camera camera, String processTitle)
+    public RegionOfInterestProcess(MainFrame mainFrame, Camera camera, String processTitle) 
+            throws Exception{
+        this(mainFrame, camera, processTitle, false);
+    }
+    
+    public RegionOfInterestProcess(MainFrame mainFrame, Camera camera, String processTitle, boolean doLocation)
             throws Exception {
         this.mainFrame = mainFrame;
         this.processTitle = processTitle;
+        this.doLocation = doLocation;
         // setup the process
         this.camera = camera;
         SwingUtilities.invokeLater(() -> {
@@ -93,43 +125,51 @@ public class RegionOfInterestProcess {
                 if (mouseLastPos != null) {
                     g2d.setColor(Color.orange);
                     g2d.setStroke(new BasicStroke(1.0f));
-                    if (step ==  0) {
+                    final ROIStep step = stepSequence.get(stepIndex);
+                    switch(step) {
+                    case UpperLeft: {
                         // draw cross hairs moving with the mouse, for the moment we're assuming a 90° aligned region of interest
                         // so these help to align with content. 
                         // TODO: perhaps rotate that with the Reticle 
                         g2d.drawLine(0, mouseLastPos.y, viewPortWidth-1, mouseLastPos.y);
                         g2d.drawLine(mouseLastPos.x, viewPortHeight-1, mouseLastPos.x, 0);
                     }
-                    else if (step == 1) {
+                    break;
+                    case UpperRight: {
                         // draw cross hairs anchored at point 0, rotating with the mouse
-                        g2d.drawLine(regionStakeout.get(0).x, regionStakeout.get(0).y, mouseLastPos.x, mouseLastPos.y);
-                        int dx = mouseLastPos.x - regionStakeout.get(0).x;
-                        int dy = mouseLastPos.y - regionStakeout.get(0).y;
+                        g2d.drawLine(regionStakeout.get(ROIStep.UpperLeft).x, regionStakeout.get(ROIStep.UpperLeft).y, mouseLastPos.x, mouseLastPos.y);
+                        int dx = mouseLastPos.x - regionStakeout.get(ROIStep.UpperLeft).x;
+                        int dy = mouseLastPos.y - regionStakeout.get(ROIStep.UpperLeft).y;
                         // draw normal on both sides
-                        g2d.drawLine(regionStakeout.get(0).x-dy, regionStakeout.get(0).y+dx, regionStakeout.get(0).x+dy, regionStakeout.get(0).y-dx);
+                        g2d.drawLine(regionStakeout.get(ROIStep.UpperLeft).x-dy, regionStakeout.get(ROIStep.UpperLeft).y+dx, regionStakeout.get(ROIStep.UpperLeft).x+dy, regionStakeout.get(ROIStep.UpperLeft).y-dx);
                         // draw normal on both sides
                         g2d.drawLine(mouseLastPos.x-dy, mouseLastPos.y+dx, mouseLastPos.x+dy, mouseLastPos.y-dx);
                     }
-                    else if (step ==  2) {
+                    break;
+                    case LowerLeft: {
                         // draw width of ROI and height both normalized and not 
-                        int dx = regionStakeout.get(1).x - regionStakeout.get(0).x;
-                        int dy = regionStakeout.get(1).y - regionStakeout.get(0).y;
+                        int dx = regionStakeout.get(ROIStep.UpperRight).x - regionStakeout.get(ROIStep.UpperLeft).x;
+                        int dy = regionStakeout.get(ROIStep.UpperRight).y - regionStakeout.get(ROIStep.UpperLeft).y;
                         // draw parallelogram
-                        g2d.drawLine(regionStakeout.get(0).x, regionStakeout.get(0).y, regionStakeout.get(1).x, regionStakeout.get(1).y);
+                        g2d.drawLine(regionStakeout.get(ROIStep.UpperLeft).x, regionStakeout.get(ROIStep.UpperLeft).y, regionStakeout.get(ROIStep.UpperRight).x, regionStakeout.get(ROIStep.UpperRight).y);
                         g2d.drawLine(mouseLastPos.x, mouseLastPos.y, mouseLastPos.x+dx, mouseLastPos.y+dy);
-                        g2d.drawLine(mouseLastPos.x, mouseLastPos.y, regionStakeout.get(0).x, regionStakeout.get(0).y);
-                        g2d.drawLine(mouseLastPos.x+dx, mouseLastPos.y+dy, regionStakeout.get(1).x, regionStakeout.get(1).y);
+                        g2d.drawLine(mouseLastPos.x, mouseLastPos.y, regionStakeout.get(ROIStep.UpperLeft).x, regionStakeout.get(ROIStep.UpperLeft).y);
+                        g2d.drawLine(mouseLastPos.x+dx, mouseLastPos.y+dy, regionStakeout.get(ROIStep.UpperRight).x, regionStakeout.get(ROIStep.UpperRight).y);
                         // also draw the normal as a guide
                         g2d.setColor(Color.red);
-                        g2d.drawLine(regionStakeout.get(0).x-dy, regionStakeout.get(0).y+dx, regionStakeout.get(0).x+dy, regionStakeout.get(0).y-dx);
+                        g2d.drawLine(regionStakeout.get(ROIStep.UpperLeft).x-dy, regionStakeout.get(ROIStep.UpperLeft).y+dx, regionStakeout.get(ROIStep.UpperLeft).x+dy, regionStakeout.get(ROIStep.UpperLeft).y-dx);
                     }
-                    else if (step ==  3) {
+                    break;
+                    case SelectMode :{
                         rectify = (mouseClickCount % 2) == 1; 
                         drawRegion(g2d);
                     }
-                    else {
+                    break;
+                    default: 
                         drawRegion(g2d);
+                        break;
                     }
+                    
                 }
                 g2d.setColor(Color.yellow);
                 g2d.setStroke(new BasicStroke(2));
@@ -141,34 +181,34 @@ public class RegionOfInterestProcess {
             protected void drawRegion(Graphics2D g2d) {
                 if (rectify) {
                     // calculate rectified region
-                    int dx1 = regionStakeout.get(1).x - regionStakeout.get(0).x;
-                    int dy1 = regionStakeout.get(1).y - regionStakeout.get(0).y;
-                    int dx2 = regionStakeout.get(2).x - regionStakeout.get(0).x;
-                    int dy2 = regionStakeout.get(2).y - regionStakeout.get(0).y;
+                    int dx1 = regionStakeout.get(ROIStep.UpperRight).x - regionStakeout.get(ROIStep.UpperLeft).x;
+                    int dy1 = regionStakeout.get(ROIStep.UpperRight).y - regionStakeout.get(ROIStep.UpperLeft).y;
+                    int dx2 = regionStakeout.get(ROIStep.LowerLeft).x - regionStakeout.get(ROIStep.UpperLeft).x;
+                    int dy2 = regionStakeout.get(ROIStep.LowerLeft).y - regionStakeout.get(ROIStep.UpperLeft).y;
                     double d = Math.sqrt((double)(dx1*dx1+dy1*dy1));
                     // normal unit vector (90° clockwise, as Y points downwards) 
                     double nx = -dy1/d;
                     double ny = dx1/d;
                     // dot product is height
                     double h = nx*dx2 + ny*dy2;
-                    int x2 = (int)Math.round(regionStakeout.get(0).x + h*nx);
-                    int y2 = (int)Math.round(regionStakeout.get(0).y + h*ny);
+                    int x2 = (int)Math.round(regionStakeout.get(ROIStep.UpperLeft).x + h*nx);
+                    int y2 = (int)Math.round(regionStakeout.get(ROIStep.UpperLeft).y + h*ny);
                     // draw rectangle
                     g2d.setColor(Color.red);
-                    g2d.drawLine(regionStakeout.get(0).x, regionStakeout.get(0).y, regionStakeout.get(1).x, regionStakeout.get(1).y);
-                    g2d.drawLine(regionStakeout.get(0).x, regionStakeout.get(0).y, x2, y2);
+                    g2d.drawLine(regionStakeout.get(ROIStep.UpperLeft).x, regionStakeout.get(ROIStep.UpperLeft).y, regionStakeout.get(ROIStep.UpperRight).x, regionStakeout.get(ROIStep.UpperRight).y);
+                    g2d.drawLine(regionStakeout.get(ROIStep.UpperLeft).x, regionStakeout.get(ROIStep.UpperLeft).y, x2, y2);
                     g2d.drawLine(x2, y2, x2+dx1, y2+dy1);
-                    g2d.drawLine(x2+dx1, y2+dy1, regionStakeout.get(1).x, regionStakeout.get(1).y);
+                    g2d.drawLine(x2+dx1, y2+dy1, regionStakeout.get(ROIStep.UpperRight).x, regionStakeout.get(ROIStep.UpperRight).y);
                     g2d.drawArc(x2-20, y2-20, 40, 40, (int)Math.round(Math.atan2(dy2, -dx2)*180/Math.PI), -90);
                 }
                 else {
                     // draw parallelogram
-                    int dx = regionStakeout.get(1).x - regionStakeout.get(0).x;
-                    int dy = regionStakeout.get(1).y - regionStakeout.get(0).y;
-                    g2d.drawLine(regionStakeout.get(0).x, regionStakeout.get(0).y, regionStakeout.get(1).x, regionStakeout.get(1).y);
-                    g2d.drawLine(regionStakeout.get(0).x, regionStakeout.get(0).y, regionStakeout.get(2).x, regionStakeout.get(2).y);
-                    g2d.drawLine(regionStakeout.get(2).x, regionStakeout.get(2).y, regionStakeout.get(2).x+dx, regionStakeout.get(2).y+dy);
-                    g2d.drawLine(regionStakeout.get(2).x+dx, regionStakeout.get(2).y+dy, regionStakeout.get(1).x, regionStakeout.get(1).y);
+                    int dx = regionStakeout.get(ROIStep.UpperRight).x - regionStakeout.get(ROIStep.UpperLeft).x;
+                    int dy = regionStakeout.get(ROIStep.UpperRight).y - regionStakeout.get(ROIStep.UpperLeft).y;
+                    g2d.drawLine(regionStakeout.get(ROIStep.UpperLeft).x, regionStakeout.get(ROIStep.UpperLeft).y, regionStakeout.get(ROIStep.UpperRight).x, regionStakeout.get(ROIStep.UpperRight).y);
+                    g2d.drawLine(regionStakeout.get(ROIStep.UpperLeft).x, regionStakeout.get(ROIStep.UpperLeft).y, regionStakeout.get(ROIStep.LowerLeft).x, regionStakeout.get(ROIStep.LowerLeft).y);
+                    g2d.drawLine(regionStakeout.get(ROIStep.LowerLeft).x, regionStakeout.get(ROIStep.LowerLeft).y, regionStakeout.get(ROIStep.LowerLeft).x+dx, regionStakeout.get(ROIStep.LowerLeft).y+dy);
+                    g2d.drawLine(regionStakeout.get(ROIStep.LowerLeft).x+dx, regionStakeout.get(ROIStep.LowerLeft).y+dy, regionStakeout.get(ROIStep.UpperRight).x, regionStakeout.get(ROIStep.UpperRight).y);
                 }
             }
         });
@@ -176,17 +216,21 @@ public class RegionOfInterestProcess {
         advance();
     }
 
+
     private MouseAdapter locationClickedListener = new MouseAdapter() {
         @Override
         public void mousePressed(final MouseEvent e) {
-            if (step < 3) {
+            switch (step) {
+            case UpperLeft:
+            case UpperRight:
+            case LowerLeft:
                 if (mouseClickCount % 2 == 0) { 
                     regionStakeout.put(step, e.getPoint());
                 }
                 else {
                     regionStakeout.remove(step);
                 }
-            }
+            };
             cameraView.flash();
             mouseClickCount++;
             cameraView.repaint();
@@ -210,26 +254,51 @@ public class RegionOfInterestProcess {
     private void advance() {
         boolean stepResult = true;
         mouseClickCount = 0;
-        step++;
-        if (step == 5) {
+        
+        if (step == ROIStep.Init) {
+            startLocation = camera.getLocation();
+        }
+
+        stepIndex += 1;
+        step = stepSequence.get(stepIndex);
+        
+        //Skip camera offset if it is not required
+        if(!doLocation && step == ROIStep.CameraOffset ) {
+            stepIndex += 1;
+            step = stepSequence.get(stepIndex);
+        }
+        
+        if (step == ROIStep.Save) {
             saveResults();
             cleanup();
-        }
-        else {
-            String title = String.format("%s (%d / 5)", processTitle, step + 1);
-            mainFrame.showInstructions(title, instructions[step], true, true,
-                    step == 4 ? "Finish" : "Next", cancelActionListener, proceedActionListener);
+        } else {
+            String title = String.format("%s (%d / %d)", processTitle, stepIndex, stepSequence.size()-1);
+            
+            mainFrame.showInstructions(title, instructions[stepIndex - 1], true, true,
+                    step == ROIStep.Complete ? "Finish" : "Next", cancelActionListener, proceedActionListener);
+            
         }
     }
 
     private boolean saveResults() {
         // calculate the Locations from pixels
         regionOfInterest = new RegionOfInterest(
-                cameraView.getCameraViewCenterOffsetsFromXy(regionStakeout.get(0).x, regionStakeout.get(0).y),
-                cameraView.getCameraViewCenterOffsetsFromXy(regionStakeout.get(1).x, regionStakeout.get(1).y),
-                cameraView.getCameraViewCenterOffsetsFromXy(regionStakeout.get(2).x, regionStakeout.get(2).y),
+                cameraView.getCameraViewCenterOffsetsFromXy(regionStakeout.get(ROIStep.UpperLeft).x, regionStakeout.get(ROIStep.UpperLeft).y),
+                cameraView.getCameraViewCenterOffsetsFromXy(regionStakeout.get(ROIStep.UpperRight).x, regionStakeout.get(ROIStep.UpperRight).y),
+                cameraView.getCameraViewCenterOffsetsFromXy(regionStakeout.get(ROIStep.LowerLeft).x, regionStakeout.get(ROIStep.LowerLeft).y),
                 rectify);
+
+        Location endLocation = camera.getLocation();
+        Location location = this.doLocation ? (endLocation.subtract(startLocation)) : null;
+        regionOfInterestOffset = new RegionOfInterestOffset(
+                cameraView.getCameraViewCenterOffsetsFromXy(regionStakeout.get(ROIStep.UpperLeft).x, regionStakeout.get(ROIStep.UpperLeft).y),
+                cameraView.getCameraViewCenterOffsetsFromXy(regionStakeout.get(ROIStep.UpperRight).x, regionStakeout.get(ROIStep.UpperRight).y),
+                cameraView.getCameraViewCenterOffsetsFromXy(regionStakeout.get(ROIStep.LowerLeft).x, regionStakeout.get(ROIStep.LowerLeft).y),
+                rectify,
+                location);
+
         UiUtils.messageBoxOnException(() -> {
+            setResult(regionOfInterestOffset);
             setResult(regionOfInterest);
         });
         return true;
@@ -255,7 +324,8 @@ public class RegionOfInterestProcess {
 
     public void setResult(RegionOfInterest roi) {
     }
-    public RegionOfInterest getRegionOfInterest() {
-        return regionOfInterest;
-    }
+
+    public void setResult(RegionOfInterestOffset roi) {
+    }    
+    
 }
